requirements.txt
fastapi
uvicorn[standard]
google-cloud-bigquery
google-cloud-storage
google-cloud-logging
yfinance
pandas
numpy
requests
python-dotenv
